// src/enums/hospital1-patient.enum.ts
export enum Hospital1Patient {
  Id = "PatientID",
  MRN = "MRN",
  DOB = "PatientDOB",
  IsAlive = "IsDeceased",
  DeathDate = "DOD_TS",
  FirstName = "FirstName",
  LastName = "LastName",
  Gender = "Gender",
  Sex = "Sex",
  Address = "Address",
  City = "City",
  State = "State",
  ZipCode = "ZipCode",
  LastModifiedDate = "LastModifiedDate",
}